Day 8

+++ src/main/kotlin/DayTemplate.kt
@@ -22,5 +22,5 @@ abstract class DayTemplate<T, G> {
     abstract fun parseInputPart2(input: List<String>): G
 
     abstract fun part1(input: T): Int
-    abstract fun part2(input: G): Int
+    abstract fun part2(input: G): Long
 }
